Add FindRecordsWithPrefix function
This function lets you efficiently process a buffer of stuffed records
that are sorted (by their decoded content).  You provide a prefix, and
we return the subset of the buffer containing all of the records whose
decoded content starts with that prefix.  (Since the records are sorted,
all of those records will be consecutive, and we can use a binary search
to find them.)